genealogy_mapper/requirements.txt
# Core dependencies
requests>=2.25.1
validators>=0.20.0
python-dateutil>=2.8.2
rich>=13.7.0
beautifulsoup4>=4.9.3
click>=8.1.7
selenium>=4.1.0
webdriver-manager>=3.5.2
neo4j>=5.17.0
openai>=1.0.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Development dependencies
pytest>=8.0.0
pytest-selenium>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
spacy>=3.7.0 